Refactor rag query pipeline
