+++ services/api/src/core/model_exceptions.py
@@ -1,0 +1,3 @@
+class ModelNotFoundError(Exception):
+    """Raised when a required model is not available and auto-pull is disabled."""
+    pass
